bug fix in IPv4 and ARP protocol and gethostbyname implementation A minor bug fixed in IPv4 checksum calculation return sum, ARP Request MAC where previously it was sending an invalid memory area for which ARP packet was destroyed. But VMware workstation somehow handled it but Virtual Box was not able to handle it.

+++ BaseHdr/Net/socket.h
@@ -95,7 +95,7 @@ typedef struct _msghdr_ {
 typedef struct _socket_ {
 	void* binedDev;
 	AuStack *rxstack;
-	int sessionPort;
+	uint16_t sessionPort;
 	int ipv4Iden;
 	int(*receive)(struct _socket_* sock, msghdr *msg, int flags);
 	int(*send)(struct _socket_* sock, msghdr* msg, int flags);
